+++ src/main/java/com/dassolt/flightfolio/service/EngineManufacturerService.java
@@ -18,6 +18,10 @@ public class EngineManufacturerService {
     }
 
     public void addEngineManufacturer(EngineManufacturer engineManufacturer) {
+        if (engineManufacturer == null || engineManufacturer.getName() == null || engineManufacturer.getName().trim().isEmpty()) {
+            throw new IllegalArgumentException("Engine manufacturer name cannot be empty");
+        }
+
         try {
             engineManufacturerDAO.add(engineManufacturer);
         } catch (SQLException e) {
@@ -26,38 +30,46 @@ public class EngineManufacturerService {
     }
 
     public EngineManufacturer retrieveEngineManufacturerById(String id) {
-        EngineManufacturer engineManufacturer = null;
+        if (id == null || id.trim().isEmpty()) {
+            throw new IllegalArgumentException("ID parameter cannot be empty.");
+        }
 
         try {
-            engineManufacturer = engineManufacturerDAO.findById(id);
+            return engineManufacturerDAO.findById(id);
         } catch (SQLException e) {
             throw new RuntimeException("Failed to retrieve engine manufacturer due to a database error", e);
         }
-
-        return engineManufacturer;
     }
 
     public List<EngineManufacturer> retrieveAllEngineManufacturers() {
-        List<EngineManufacturer> engineManufacturers = null;
-
         try {
-            engineManufacturers = engineManufacturerDAO.findAll();
+            return engineManufacturerDAO.findAll();
         } catch (SQLException e) {
             throw new RuntimeException("Failed to retrieve all engine manufacturers due to a database error", e);
         }
-
-        return engineManufacturers;
     }
 
     public void updateEngineManufacturer(EngineManufacturer engineManufacturer) {
+        if (engineManufacturer == null || engineManufacturer.getId() == null) {
+            throw new IllegalArgumentException("Engine manufacturer and its ID cannot be null.");
+        }
+
         try {
-            engineManufacturerDAO.update(engineManufacturer);
+            EngineManufacturer engineManufacturerToUpdate = engineManufacturerDAO.findById(engineManufacturer.getId());
+
+            if (engineManufacturerToUpdate != null && !engineManufacturerToUpdate.equals(engineManufacturer)) {
+                engineManufacturerDAO.update(engineManufacturer);
+            }
         } catch (SQLException e) {
             throw new RuntimeException("Failed to update engine manufacturer due to a database error", e);
         }
     }
 
     public void deleteEngineManufacturer(EngineManufacturer engineManufacturer) {
+        if (engineManufacturer == null || engineManufacturer.getId() == null) {
+            throw new IllegalArgumentException("Engine manufacturer and its ID cannot be null.");
+        }
+
         try {
             engineManufacturerDAO.delete(engineManufacturer);
         } catch (SQLException e) {
