+++ src/main/java/com/dassolt/flightfolio/service/ManufacturerService.java
@@ -18,6 +18,10 @@ public class ManufacturerService {
     }
 
     public void addManufacturer(Manufacturer manufacturer) {
+        if (manufacturer == null || manufacturer.getName() == null || manufacturer.getName().trim().isEmpty()) {
+            throw new IllegalArgumentException("Manufacturer name cannot be empty");
+        }
+
         try {
             manufacturerDAO.add(manufacturer);
         } catch (SQLException e) {
@@ -26,38 +30,42 @@ public class ManufacturerService {
     }
 
     public Manufacturer retrieveManufacturerById(String id) {
-        Manufacturer manufacturer = null;
+        if (id == null || id.trim().isEmpty()) {
+            throw new IllegalArgumentException("ID parameter cannot be empty.");
+        }
 
         try {
-            manufacturer = manufacturerDAO.findById(id);
+            return manufacturerDAO.findById(id);
         } catch (SQLException e) {
             throw new RuntimeException("Failed to retrieve manufacturer due to a database error", e);
         }
-
-        return manufacturer;
     }
 
     public List<Manufacturer> retrieveAllManufacturers() {
-        List<Manufacturer> manufacturers = null;
-
         try {
-            manufacturers = manufacturerDAO.findAll();
+            return manufacturerDAO.findAll();
         } catch (SQLException e) {
             throw new RuntimeException("Failed to retrieve all manufacturers due to a database error", e);
         }
-
-        return manufacturers;
     }
 
     public void updateManufacturer(Manufacturer manufacturer) {
         try {
-            manufacturerDAO.update(manufacturer);
+            Manufacturer manufacturerToUpdate = manufacturerDAO.findById(manufacturer.getId());
+
+            if (manufacturerToUpdate != null && !manufacturerToUpdate.equals(manufacturer)) {
+                manufacturerDAO.update(manufacturer);
+            }
         } catch (SQLException e) {
             throw new RuntimeException("Failed to update manufacturer due to a database error", e);
         }
     }
 
     public void deleteManufacturer(Manufacturer manufacturer) {
+        if (manufacturer == null || manufacturer.getId() == null) {
+            throw new IllegalArgumentException("Manufacturer and its ID cannot be null.");
+        }
+
         try {
             manufacturerDAO.delete(manufacturer);
         } catch (SQLException e) {
